删除不需要的js css文件 添加model验证

+++ db/migrate/20190209234533_create_citrines.rb
@@ -1,8 +1,8 @@
 class CreateCitrines < ActiveRecord::Migration
   def change
     create_table :citrines do |t|
-      t.integer :count, null: false, default: 0
-      t.integer :total, null: false, default: 0
+      t.float :count, null: false, default: 0
+      t.float :total, null: false, default: 0
       t.references :user
 
       t.timestamps null: false
